remove document_source.rs
